// src/app/enums/colors.enum.ts
export enum ColorsEnum {
  fairPink = '#ffeded',
  blazeOrange = '#ff6600',
  errorLight = '#FF5733',
  errorDark = '#FEFEFE',
  warningLight = '#f8fdd7',
  warningDark = '#f0ad4e',
  infoLight = '#befbff',
  infoDark = '#5bc0de',
  successLight = '#ddfdd1',
  successDark = '#5cb85c',
  purple = '#b23deb',
  lightBlue = '#73d3fa',
  red = '#f44336',
  orange = '#ffa240',
  black = '#0d0d0d',
  mediumGray = '#8b8b8b'
}